Specify cert bundle path for superset oauth
We've found that in specific configurations of openshift, it is
necessary to overrided the value of the default python requests
library's certificate bundle location. This change forces the oauth call
here to use the cert bundle that we create up above, which should allow
oauth to keep working when customized deployments have overriden the
value of the "CURL_CA_BUNDLE" env var.